DOC: Inline annotation user demos
`annotate_explain.py` and `connectionstyle_demo.py` can be inlined where
they are used in the Annotations tutorial.

`custom_boxstyle01.py` is half written in the "Defining custom box
styles" section, so just insert the second half.

`annotate_text_arrow.py` is better explained in the "Annotating with
boxed text" section of the Annotations tutorial.

`simple_annotate01.py` repeats all of the tutorial, but without any
explanatory text.
